+++ lib/src/custom_refresh_indicator.dart
@@ -379,6 +379,8 @@ class CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   }
 
   bool _handleOverscrollNotification(OverscrollNotification notification) {
+    controller.setIndicatorDragDetails(notification.dragDetails);
+
     if (!controller.hasEdge) {
       controller.setIndicatorEdge(
         notification.overscroll.isNegative
@@ -399,7 +401,9 @@ class CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   }
 
   bool _handleScrollEndNotification(ScrollEndNotification notification) {
-    controller.clearPhysicsState();
+    controller
+      ..setIndicatorDragDetails(null)
+      ..clearPhysicsState();
 
     if (controller.state.isArmed) {
       _start();